requirements.txt
aiomysql==0.2.0
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
click==8.1.7
databases==0.8.0
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
mysql-connector-python==8.2.0
passlib==1.7.4
protobuf==4.21.12
pyasn1==0.5.1
pydantic==2.5.3
pydantic_core==2.14.6
PyMySQL==1.1.0
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.51
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
